requirements.txt
--constraint "https://raw.githubusercontent.com/apache/airflow/constraints-2.8.1/constraints-3.11.txt"
apache-airflow[]==2.8.1
apache-airflow-providers-postgres==5.10.0
apache-airflow-providers-sftp==4.8.1
apache-airflow-providers-apache-spark==4.7.0
black==24.1a1
pytest==7.4.4
flake8==7.0.0
isort==5.13.2
pre-commit-hooks==v4.5.0
Faker==22.6.0
pytest-mock==3.12.0
pyspark==3.5.0
pandas==2.1.4
pyarrow==14.0.2
